requirements.txt
factorio-rcon-py[async]==2.0.1
discord==2.3.2
